==> 01-elecciones_municipales.Rmd <==
---
title: "Merge datos municipales"
output: html_document
date: "2023-10-30"
editor_options: 
  chunk_output_type: console
---

Los paquetes que utilizaremos:

```{r warning=F, message=F}
library(tidyverse)
library(stringi)
library(janitor)
library(lubridate)
library(sjmisc)
```

# Base 1992-2016

Revisar votos válidamente emitidos para el 2016 = Filtrar los nominados

Carga de base de datos con las elecciones de 1992 a 2016:

```{r}
df_elections_1992_2016 <- readxl::read_excel("input/municipales_1992_2016.xlsx") %>% 
  mutate(fecha = year(fecha_eleccion)) %>% 
  select(-fecha_eleccion) %>% 
  mutate(candidato = str_to_upper(candidato)) %>% 
  filter(!candidato %in% c("TOTAL VOTACIÓN", "VÁLIDAMENTE EMITIDOS"))

colnames(df_elections_1992_2016)

df_elections_1992_2016 <- df_elections_1992_2016 %>% 
  select(tipo_eleccion, fecha, cargo, comuna, candidato, sexo, electo, lista, partido, sigla_partido, votos)
```

Carga de la base de datos 2021:

```{r}
df_eleccions_2021 <- read_rds("output/231029 - BBDD Alcaldes 2021.rds") %>% 
  mutate(tipo_eleccion = "MUNICIPALES",
         cargo = "ALCALDE",
         fecha = 2021)

df_eleccions_2021 <- df_eleccions_2021 %>% 
  mutate(comuna = str_replace(comuna, "LLAILLAY", "LLAY-LLAY"),
         comuna = str_replace(comuna, "AYSEN", "AISEN"),
         comuna = str_replace(comuna, "TREHUACO", "TREGUACO")) %>% 
  filter(!(candidato == "GARY LEIVA ROJO" & is.na(pacto)))

colnames(df_eleccions_2021)
```

Unir ambas bases de datos:

```{r}
df_county_elections <- bind_rows(df_elections_1992_2016, df_eleccions_2021)

rm(df_elections_1992_2016, df_eleccions_2021)

colnames(df_county_elections)
```

Ordenar la base de datos:

```{r}
df_county_elections <- df_county_elections %>% 
  select(tipo_eleccion, fecha, cargo, comuna, candidato, sexo, electo, lista, pacto, partido,
         sigla_partido, votos, votos_candidatos) %>% 
  mutate(votos = coalesce(votos, votos_candidatos)) %>% 
  select(-votos_candidatos) %>% 
  filter(!(cargo %in% c("CONCEJAL","NOMINADO") & fecha > 2001))

#writexl::write_xlsx(df_county_elections, "output/elecciones municipales 1992 2021.xlsx")

frq(df_county_elections, fecha)

df_county_elections_rank <- df_county_elections %>% 
  group_by(fecha, comuna) %>% 
  mutate(votos_comunales = sum(votos)) %>% 
  filter(!candidato %in% c("VOTOS NULOS","VOTOS EN BLANCO")) %>% 
  arrange(votos, .by_group = T) %>% 
  mutate(rank_votos = as.integer(rank(-votos))) %>% 
  group_by(fecha, comuna, lista, pacto)
```

## Base de datos votos nulos, blanco, válidamente emitidos y totales.

```{r}
df_votes <- df_county_elections %>% 
  select(fecha, cargo, comuna, candidato, votos) %>% 
  filter(candidato %in% c("VOTOS NULOS","VOTOS EN BLANCO")) %>%
  pivot_wider(names_from = candidato, values_from = votos) %>% 
  clean_names()
```

```{r}
df_elections_votes <- left_join(df_county_elections_rank, df_votes)

df_elections_votes <- df_elections_votes %>% 
  rowwise() %>% 
  mutate(votos_validos = votos_comunales - sum(votos_nulos, votos_en_blanco))

rm(df_county_elections, df_county_elections_rank, df_votes)

df_elections_votes <- df_elections_votes %>% 
  ungroup() %>% 
  mutate(pacto = coalesce(lista, pacto)) %>% 
  select(fecha, tipo_eleccion, cargo, comuna, candidato, sexo, electo, rank_votos, pacto, partido, sigla_partido, 
         votos_candidato = votos, votos_nulos, votos_en_blanco, votos_comunales, votos_validos)
```

Guardamos esa base:

```{r eval=F}
# write_rds(df_completa_votacion, "output/231030 - Base completa elecciones municipales.rds")
# haven::write_dta(df_completa_votacion, "output/231030 - Base completa elecciones municipales.dta")
```

## Base de listas y partidos políticos por elección:

Revisión de partidos:

```{r}
df_elections_votes %>% 
  ungroup() %>% 
  frq(partido)
```

```{r}
df_list <- df_elections_votes %>% 
  ungroup() %>% 
  mutate_at(vars(pacto, partido), ~str_to_upper(.), ~str_squish(.)) %>% 
  select(fecha, pacto, partido) %>% 
  distinct() %>% 
  arrange(fecha, pacto)
```

Guardamos la base de datos con la información de partidos, listas y pactos:

```{r eval = F}
#writexl::write_xlsx(df_list, "output/231030 - Pactos.xlsx")
```

Leemos la base de datos de pactos/listas electorales actualizada y estandarizada:

```{r}
df_list <- readxl::read_xlsx("input/231030 - Pactos_rec.xlsx") %>% 
  rename(pacto = lista, pacto_nombre = pacto_rec, coalicion = categoria)
```

Unimos y reemplazamos las columnas por las nuevas:

```{r}
df_elections_votes <- df_elections_votes %>% 
  left_join(df_list) %>% 
  select(-partido) %>% 
  rename(partido = partido_rec)
```

## Votos

Calculamos los votos obtenidos por pacto y luego por partido:

```{r}
df_municipal_elections <- df_elections_votes %>% 
  group_by(fecha, comuna, pacto_nombre) %>%
  mutate(votos_lista = sum(votos_candidato, na.rm = T)) %>%
  group_by(fecha, comuna, partido) %>%
  mutate(votos_partido = sum(votos_candidato, na.rm = T)) %>%
  group_by(fecha, comuna, coalicion) %>% 
  mutate(votos_coalicion = sum(votos_candidato, na.rm = T)) %>% 
  ungroup()

df_n_candidates <- df_elections_votes %>% 
  group_by(fecha, comuna, coalicion) %>% 
  mutate(n_candidatos = n()) %>% 
  select(fecha, comuna, coalicion, n_candidatos) %>% 
  unique() %>% 
  mutate(coalicion = str_to_lower(coalicion)) %>% 
  pivot_wider(names_from = coalicion, 
              values_from = n_candidatos, 
              names_glue = "{coalicion}_{.value}") %>% 
  janitor::clean_names() %>% select(-8)

df_municipal_elections <- left_join(df_municipal_elections, df_n_candidates)

rm(df_list, df_elections_votes, df_n_candidates)
```


```{r}
colnames(df_municipal_elections)

df_coalition <- df_municipal_elections %>% 
  select(fecha, cargo, comuna, coalicion, votos_coalicion) %>%
  distinct() %>% 
  pivot_wider(values_from = votos_coalicion, names_from = coalicion, names_prefix = "coalicion") %>% 
  clean_names() %>% select(-coalicion_na)
```

```{r}
df_municipal_elections <- df_municipal_elections %>% 
  select(-sigla_partido) %>% 
  left_join(df_coalition)

rm(df_coalition)
```

Pasamos a %

```{r}
df_municipal_elections <- df_municipal_elections %>% 
  rowwise() %>% 
  mutate(coalicion_concertacion_p = coalicion_concertacion/votos_validos,
         coalicion_alianza_p = coalicion_alianza/votos_validos,
         coalicion_independientes_p = coalicion_independientes/votos_validos,
         coalicion_otros_p = coalicion_otros/votos_validos,
         coalicion_frente_amplio_p = coalicion_frente_amplio/votos_validos)
```


Vemos las columnas que tenemos hasta el momento:

```{r}
colnames(df_municipal_elections)

df_all_candidates_by_year <- df_municipal_elections %>% 
  mutate(votos_candidato_p = votos_candidato/votos_validos) %>% 
  select(1:6, 9:11, pacto_nombre:coalicion, votos_candidato_p)
```

Ahora, dejamos a la primera y segunda mayoría por comuna. Seleccionamos `ALCALDE` en las elecciones de 1992. 1996 y 2000.

```{r}
df_municipal_elections <- df_municipal_elections %>% 
  filter(rank_votos %in% c(1,2) | electo == "ALCALDE") %>% 
  ungroup()
```

Seleccionando de esta manera, quedan comunas donde la segunda mayoría es "VOTO EN BLANCO" o "VOTO NULO".

Nos quedan 3094 observaciones.

```{r}
flat_table(df_municipal_elections, coalicion, fecha)
```

Revisamos si coincide los alcaldes con las votaciones en los años 1992, 1996 y 2000:

```{r}
flat_table(df_municipal_elections, rank_votos)
```

Revisamos los alcaldes donde no tienen la primera mayoría en su comuna:

```{r}
df_municipal_elections %>% 
  select(fecha, comuna, candidato, electo, rank_votos) %>% 
  filter(rank_votos > 2) %>% 
  print(n = 100)
```

>Todos están en las elecciones de 1992, por lo que es posible que haya sido electo alcalde sin tener la primera mayoría. En las comunas con más de 2 alcaldes por periodo (1992), queda registrado el último alcalde que estuvo en el cargo.

## Género

Para las últimas elecciones, no existe la columna género. Por ese motivo, tendremos que digitarlo:

```{r eval = F}
# df_municipal_elections %>%
#   filter(fecha == 2021) %>%
#   select(fecha, comuna, candidato, rank_votos) %>%
#   writexl::write_xlsx("output/231030 - Género.xlsx")
```

```{r}
df_municipal_elections <- df_municipal_elections %>% 
  filter(!is.na(comuna))
```

Leemos la base de datos nueva con el género de los candidatos de las elecciones 2021:

```{r}
df_gender <- readxl::read_excel("input/231030 - Género_rec.xlsx")
```

```{r}
df_municipal_elections <- df_municipal_elections %>%
  left_join(df_gender)

df_municipal_elections <- df_municipal_elections %>% 
  mutate(sexo = coalesce(sexo, genero)) %>% 
  select(-genero)

rm(df_gender)
```

```{r}
flat_table(df_municipal_elections, fecha, sexo)

df_municipal_elections %>% 
  filter(is.na(sexo)) %>% 
  select(comuna, candidato)

df_gender <- df_municipal_elections %>% 
  select(comuna, candidato, fecha, sexo)

#writexl::write_xlsx(df_gender, "output/231129 - Género.xlsx")
```

Revisamos si hay candidatos sin sexo asignado:

Hay datos sin sexo asignado, así que volvemos a hacer la revisión con una base de datos manual:

```{r}
df_gender <- readxl::read_xlsx("input/231129 - Género_rev.xlsx") %>% 
  unique()

df_municipal_elections <- df_municipal_elections %>% 
  distinct()

df_municipal_elections <- df_municipal_elections %>% 
  select(-sexo) %>% 
  left_join(df_gender, by = join_by(comuna == comuna, fecha == fecha, candidato == candidato)) %>% 
  group_by(fecha, comuna, candidato) %>% 
  slice(., 1)

df_municipal_elections %>% 
  filter(is.na(sexo)) %>% 
  select(comuna, fecha, candidato)
```

Seleccionamos las variables de interés:

```{r}
df_winner <- df_municipal_elections %>% 
  mutate(
  puestos_votacion = case_when(
    electo == "ALCALDE" ~ 'primera_mayoria',
    rank_votos == 1 & fecha == 2021 ~ 'primera_mayoria',
    rank_votos == 2 ~ 'segunda_mayoria',
    T ~ "otro")) %>% 
  filter(puestos_votacion == 'primera_mayoria')

df_runner_up <- df_municipal_elections %>% 
  mutate(
  puestos_votacion = case_when(
    electo == "ALCALDE" ~ 'primera_mayoria',
    rank_votos == 1 & fecha == 2021 ~ 'primera_mayoria',
    rank_votos == 2 ~ 'segunda_mayoria',
    T ~ "otro")) %>% 
  filter(puestos_votacion != 'primera_mayoria')

```

```{r}
colnames(df_winner)

df_winner <- df_winner %>% 
  rename_at(vars(candidato, sexo, pacto, pacto_nombre, partido, votos_candidato, coalicion), 
            ~paste0(., "_pm")) %>% 
  select(-puestos_votacion, -rank_votos)

df_runner_up <- df_runner_up %>% 
  rename_at(vars(candidato, sexo, pacto, pacto_nombre, partido, votos_candidato, coalicion), 
            ~paste0(., "_sm"))

df_runner_up <- df_runner_up %>% 
  group_by(fecha, comuna) %>% 
  unique() %>% 
  mutate(mas_de_uno = if_else(length(duplicated(comuna))==2, TRUE, FALSE)) %>% 
  group_by(fecha, comuna) %>% 
  slice_max(votos_candidato_sm) %>% 
  select(-mas_de_uno, -puestos_votacion, -rank_votos)

colnames(df_runner_up)
```


```{r}
df_municipal_elections <- left_join(df_winner, df_runner_up, by = join_by(fecha, comuna))

df_municipal_elections <- df_municipal_elections %>% 
  select(-ends_with(".y")) %>% 
  rename_at(vars(ends_with(".x")), ~str_remove(., ".x"))

rm(df_runner_up, df_winner)
```

Calculamos el % a las columnas de `primera_votacion` y `segunda_votacion`, además de crear la `dummy_genero` y `brecha_victoria`.

```{r}
df_municipal_elections <- df_municipal_elections %>% 
  rowwise() %>% 
  mutate(votos_candidato_pm_p = votos_candidato_pm/votos_validos,
         votos_candidato_sm_p = votos_candidato_sm/votos_validos,
         brecha_victoria = votos_candidato_pm - votos_candidato_sm,
         brecha_victoria_p = votos_candidato_pm_p - votos_candidato_sm_p) %>% 
  select(-electo)
```

## Coalición

Ahora, haremos las variables relacionadas con la reeleccion de coalicion incumbente:

```{r}
df_municipal_incumbent <- df_municipal_elections %>% 
  group_by(comuna) %>% 
  mutate(reeleccion_coalicion_alcalde = if_else(lag(coalicion_pm) == coalicion_pm, 1, 0),
         porc_coalicion_alcalde_electa = case_when(
           coalicion_pm == "CONCERTACION"   ~ coalicion_concertacion_p,
           coalicion_pm == "ALIANZA"        ~ coalicion_alianza_p,
           coalicion_pm == "INDEPENDIENTES" ~ coalicion_independientes_p,
           coalicion_pm == "OTROS"          ~ coalicion_otros_p,
           coalicion_pm == "FRENTE AMPLIO"  ~ coalicion_frente_amplio_p))

# Revisión por comuna

df_municipal_incumbent %>% 
  select(fecha, comuna, candidato_pm, reeleccion_coalicion_alcalde, porc_coalicion_alcalde_electa) %>% 
  filter(comuna == "SANTIAGO") 

# Revisión de conteo

df_municipal_incumbent <- df_municipal_incumbent %>% 
  #select(fecha, comuna, candidato_pm, reeleccion_coalicion_alcalde) %>% 
  mutate(reeleccion_coalicion_alcalde = replace_na(reeleccion_coalicion_alcalde, value = 0)) %>% 
  mutate(periodos_consecutivos_coalicion = hutilscpp::cumsum_reset(reeleccion_coalicion_alcalde >= 1,
                                                                 reeleccion_coalicion_alcalde))

df_municipal_incumbent %>% 
  select(fecha, comuna, candidato_pm, reeleccion_coalicion_alcalde, periodos_consecutivos_coalicion) %>% 
  filter(comuna == "TALTAL")
```

Ahora, reeleccion por candidato:

Primero, haremos una nueva base de datos con información del año, comuna y nombre del candidato:

```{r}
df_winner <- df_municipal_incumbent %>% 
  select(fecha, comuna, candidato_pm, reeleccion_coalicion_alcalde, periodos_consecutivos_coalicion) %>% 
  arrange(comuna, fecha)
```

Guardamos la base de datos para su revisión manual:

```{r}
#writexl::write_xlsx(df_winner, "output/231030 - Candidatos.xlsx")
```

## Reeleción de candidatos

Cargamos la base de datos con la reelección de candidatos digitada:

```{r}
df_winner <- readxl::read_excel("input/231030 - Candidatos_rec.xlsx") %>% 
  select(1,2,3,7) %>% 
  rename(periodos_consecutivos_alcalde = 4) %>% 
  mutate(reeleccion_alcalde = if_else(periodos_consecutivos_alcalde != 1, 1, 0))
```

Unimos la base de datos con la planilla de reelección:

```{r}
df_municipal_incumbent <- df_municipal_incumbent %>% 
  left_join(df_winner, by = join_by(fecha, comuna, candidato_pm))

df_municipal_incumbent <- df_municipal_incumbent %>% 
  group_by(comuna) %>% 
  arrange(fecha, .by_group = T)

df_municipal_incumbent <- df_municipal_incumbent %>% 
  mutate(periodos_consecutivos_coalicion = periodos_consecutivos_coalicion + 1,
         coalicion_municipal_incumbente = lag(coalicion_pm), 
         porc_coalicion_alcalde_incumbente = case_when(
           coalicion_municipal_incumbente == "CONCERTACION" ~ coalicion_concertacion_p,
           coalicion_municipal_incumbente == "ALIANZA" ~ coalicion_alianza_p,
           coalicion_municipal_incumbente == "INDEPENDIENTES" ~ coalicion_independientes_p,
           coalicion_municipal_incumbente == "OTROS" ~ coalicion_otros_p,
           coalicion_municipal_incumbente == "FRENTE AMPLIO" ~ coalicion_frente_amplio_p
         ))

df_municipal_incumbent %>% 
  filter(comuna == "TALTAL") %>% 
  select(fecha, comuna, candidato_pm, periodos_consecutivos_alcalde)

df_municipal_incumbent %>% 
  select(fecha, comuna, candidato_pm, coalicion_pm, coalicion_municipal_incumbente, porc_coalicion_alcalde_incumbente)
```

Transformamos las variables de chr a minúscula:

```{r}
df_municipal_incumbent <- df_municipal_incumbent %>% 
  select(tipo_eleccion, cargo, comuna, candidato_pm, candidato_sm, sexo_pm, sexo_sm, pacto_nombre_pm, pacto_nombre_sm,
         partido_pm, partido_sm, coalicion_pm, coalicion_sm, everything()) %>% 
  mutate(across(where(is.character), 
                \(x) stringi::stri_trans_general(str_to_lower(x), "latin-ascii"))) %>% 
  mutate(comuna = stringi::stri_trans_general(str_to_lower(comuna), "latin-ascii"))

df_municipal_incumbent <- df_municipal_incumbent %>% ungroup()
```

Agregamos la variable relativa a si el alcalde incumbente se presenta o no a la reelección:

```{r}
df_electoral_participation <- haven::read_dta("input/base_consolidada_promedios_2001_2019.dta") %>% 
  select(municipio, cod_com, candidato, periodo_electoral, participacion_alcalde_incum, elecciones_no_competitivas) %>% 
  mutate(fecha = case_when(
    periodo_electoral == 1 ~ 2004,
    periodo_electoral == 2 ~ 2008,
    periodo_electoral == 3 ~ 2012,
    periodo_electoral == 4 ~ 2016
  ))

```

Unimos las bases de datos utilizando

```{r}

df_municipal_incumbent <- df_municipal_incumbent %>% 
  filter(!comuna %in% c("cabo de hornos","cabo de hornos y antartica", "cabo de hornos(ex-navarino)", "antartica")) %>% 
  mutate(comuna = str_replace_all(comuna, 
                                  c("llay-llay" = "llaillay","treguaco" = "trehuaco", "o'higgins" = "o´higgins",
                                    "marchigue" = "marchihue", "aisen" = "aysen")))

df_incumbent_participation <- df_municipal_incumbent %>% 
  left_join(df_electoral_participation, by = join_by(comuna == municipio, candidato_pm == candidato, fecha)) %>% 
  group_by(fecha, comuna) %>% 
  slice(., 1)

df_incumbent_participation <- df_incumbent_participation %>% 
  group_by(comuna) %>% 
  fill(., cod_com, .direction = "downup") %>% 
  mutate(periodo_electoral = seq_along(periodo_electoral))

df_incumbent_participation %>% 
  filter(comuna == "rio claro") %>% 
  select(fecha, comuna, periodo_electoral, participacion_alcalde_incum, elecciones_no_competitivas)

df_incumbent_participation %>% 
  filter(comuna == "yungay") %>% 
  select(fecha, comuna, candidato_pm, candidato_sm, votos_candidato_pm, votos_candidato_sm)

# rio claro: se repite 4 veces 1996
# san juan de la costa: se repite 4 veces el 2021
# tortel: se repite 4 veces el 2021
# vilcun: se repite 4 veces el 2021
# yungay: se repite 4 veces el 2021

# writexl::write_xlsx(df_incumbent_participation, "output/231112 - Participación.xlsx")

```

```{r}

df_electoral_participation <- readxl::read_excel("input/231115 - Participación_rec.xlsx") %>%
  select(comuna, cod_com, fecha, participacion_alcalde_incum, elecciones_no_competitivas) %>% 
  mutate(elecciones_no_competitivas = replace_na(elecciones_no_competitivas, value = 0))

df_electoral_participation %>% 
  filter(comuna == "rio claro")

df_incumbent_participation <- df_incumbent_participation %>% 
  group_by(fecha, comuna) %>% 
  slice(., 1) %>% 
  select(-c(participacion_alcalde_incum, elecciones_no_competitivas, cod_com))

colnames(df_incumbent_participation)

df_incumbent_participation <- df_incumbent_participation %>% left_join(df_electoral_participation,
                                                               by = join_by(comuna == comuna, 
                                                                            fecha == fecha))

df_incumbent_participation <- df_incumbent_participation %>% 
  group_by(comuna) %>% 
  fill(., cod_com, .direction = "downup") %>% 
  mutate(periodo_electoral = seq_along(fecha))

df_incumbent_participation %>% 
  filter(comuna == "rio claro") %>% 
  select(fecha, comuna, periodo_electoral, participacion_alcalde_incum, elecciones_no_competitivas)

df_incumbent_participation <- df_incumbent_participation %>% 
  filter(!periodo_electoral %in% c(9, 10, 11))
```

Revisión de variables y ubicación:

```{r}
colnames(df_municipal_incumbent)
```

# Variables nuevas: 

- Información (primeras 7 columnas) para el incumbente. Tener en consideración la participación, no la reelección.

```{r incumbente}
df_incumbent <- df_incumbent_participation %>% 
  select(fecha, comuna, candidato_pm, sexo_pm, pacto_nombre_pm, pacto_pm, partido_pm, 
         coalicion_pm, reeleccion_alcalde, participacion_alcalde_incum, votos_candidato_pm,
         votos_candidato_pm_p) %>% 
  mutate(incumbente = case_when(
    reeleccion_alcalde == 1 & participacion_alcalde_incum == 1 ~ lag(candidato_pm),
    reeleccion_alcalde == 0 & participacion_alcalde_incum == 1 ~ lag(candidato_pm),
    reeleccion_alcalde == 0 & participacion_alcalde_incum == 0 ~ NA_character_
  )) %>% filter(fecha > 2000)

df_all_candidates_by_year <- df_all_candidates_by_year %>% 
    mutate(across(where(is.character), 
                \(x) stringi::stri_trans_general(str_to_lower(x), "latin-ascii"))) %>% 
    filter(!comuna %in% c("cabo de hornos","cabo de hornos y antartica", "cabo de hornos(ex-navarino)", "antartica")) %>% 
    mutate(comuna = str_replace_all(comuna, 
                                  c("llay-llay" = "llaillay","treguaco" = "trehuaco", "o'higgins" = "o´higgins",
                                    "marchigue" = "marchihue", "aisen" = "aysen")))

colnames(df_all_candidates_by_year)

df_all_candidates_by_year <- df_all_candidates_by_year %>% 
  filter(fecha > 2000) %>% 
  select(-tipo_eleccion, -cargo, -votos_nulos, sexo_incumbente = sexo, pacto_incumbente = pacto, 
         votos_incumbente = votos_candidato, 
         pacto_nombre_incumbente = pacto_nombre, partido_incumbente = partido,
         coalicion_incumbente = coalicion, votos_incumbente_p = votos_candidato_p)

#writexl::write_xlsx(df_all_candidates_by_year, "all candidates.xlsx")

df_join <- left_join(df_incumbent, df_all_candidates_by_year, by = join_by("fecha" == "fecha", 
                                                                           "comuna" == "comuna", 
                                                                           "incumbente" == "candidato",))

df_elect_and_incumbent <- df_join %>% 
  mutate(coalicion_incumbente = if_else(reeleccion_alcalde == 1, coalicion_pm, coalicion_incumbente),
         sexo_incumbente = if_else(reeleccion_alcalde == 1, sexo_pm, sexo_incumbente),
         pacto_incumbente = if_else(reeleccion_alcalde == 1, pacto_pm, pacto_incumbente),
         pacto_nombre_incumbente = if_else(reeleccion_alcalde == 1, pacto_nombre_pm, pacto_nombre_incumbente),
         partido_incumbente = if_else(reeleccion_alcalde == 1, partido_pm, partido_incumbente),
         votos_incumbente = if_else(reeleccion_alcalde == 1, votos_candidato_pm, votos_incumbente),
         votos_incumbente_p = if_else(reeleccion_alcalde == 1, votos_candidato_pm_p, votos_incumbente_p))

#writexl::write_xlsx(df_elect_and_incumbent, "output/electo e incumbente.xlsx")

df_elect_and_incumbent2 <- readxl::read_xlsx("input/electo e incumbente_rev.xlsx") %>% 
  select(fecha, comuna, incumbente:votos_incumbente_p)

colnames(df_elect_and_incumbent2)

df_elect_and_incumbent2 %>% 
  filter(comuna == "santiago") %>% 
  select(fecha, comuna, incumbente, sexo_incumbente, votos_incumbente_p)

df_incumbent_participation <- df_incumbent_participation %>% 
  left_join(df_elect_and_incumbent2)

df_incumbent_participation %>% 
  filter(comuna == "santiago") %>% 
  select(fecha, comuna, candidato_pm, sexo_pm, incumbente, sexo_incumbente)
```


- Arreglar la columna con cantidad de periodos electorales. Parte en cero, luego en 1 siempre.

```{r}
df_incumbent_participation <- df_incumbent_participation %>% 
  group_by(comuna) %>% 
  mutate(periodos_consecutivos_alcalde = case_when(
    fecha == 1992 ~ 0,
    T ~ periodos_consecutivos_alcalde
  )) %>% 
  mutate(periodos_consecutivos_alcalde = lag(periodos_consecutivos_alcalde)) %>% 
  mutate(periodos_consecutivos_alcalde = case_when(    
    fecha == 1992 ~ 0,
    fecha == 1996 ~ 1,
    T ~ periodos_consecutivos_alcalde)) %>% 
  mutate(periodos_consecutivos_coalicion = case_when(
    fecha == 1992 ~ 0,
    T ~ periodos_consecutivos_coalicion
  )) %>% 
  mutate(periodos_consecutivos_coalicion = lag(periodos_consecutivos_coalicion)) %>% 
  mutate(periodos_consecutivos_coalicion = case_when(    
    fecha == 1992 ~ 0,
    fecha == 1996 ~ 1,
    T ~ periodos_consecutivos_coalicion))
```

- % de la coalicion incumbente

```{r}

df_incumbent_participation <- df_incumbent_participation %>% 
  rowwise() %>% 
  mutate(votos_coalicion_p = votos_coalicion/votos_validos) %>% 
  group_by(comuna) %>% 
  mutate(votos_coalicion_incumbente_p = case_when(
    independientes_n_candidatos == 1 & coalicion_municipal_incumbente == "independientes" ~ coalicion_independientes_p,
    independientes_n_candidatos != 1 & coalicion_municipal_incumbente == "independientes" ~ votos_candidato_pm_p,
    concertacion_n_candidatos == 1 & coalicion_municipal_incumbente == "concertacion" ~ coalicion_concertacion_p,
    concertacion_n_candidatos != 1 & coalicion_municipal_incumbente == "concertacion" ~ votos_candidato_pm_p,
    alianza_n_candidatos == 1 & coalicion_municipal_incumbente == "alianza" ~ coalicion_alianza_p,
    alianza_n_candidatos != 1 & coalicion_municipal_incumbente == "alianza" ~ votos_candidato_pm_p,
    otros_n_candidatos == 1 & coalicion_municipal_incumbente == "otros" ~ coalicion_otros_p,
    otros_n_candidatos != 1 & coalicion_municipal_incumbente == "otros" ~ votos_candidato_pm_p,
    frente_amplio_n_candidatos == 1 & coalicion_municipal_incumbente == "frente amplio" ~ coalicion_frente_amplio_p,
    frente_amplio_n_candidatos != 1 & coalicion_municipal_incumbente == "frente amplio" ~ votos_candidato_pm_p
  ))

```

```{r}
colnames(df_incumbent_participation)
```

# ENP

```{r}
df_enp <- read_rds("data/enp.rds") %>% 
  select(fecha, comuna, enp_candidatos, enp_coalicion, enp_partido, enp_pacto, candidatos_n, coaliciones_n, partidos_n)

colnames(df_enp)

df_enp <- df_enp %>% 
    mutate(comuna = stringi::stri_trans_general(str_to_lower(comuna), "latin-ascii")) %>% 
    filter(!comuna %in% c("cabo de hornos","cabo de hornos y antartica", "cabo de hornos(ex-navarino)", "antartica")) %>% 
    mutate(comuna = str_replace_all(comuna, 
                                  c("llay-llay" = "llaillay","treguaco" = "trehuaco", "o'higgins" = "o´higgins",
                                    "marchigue" = "marchihue", "aisen" = "aysen")))

df_incumbent_participation <- df_incumbent_participation %>% 
  left_join(df_enp, by = join_by(fecha == fecha, comuna == comuna))
```

74 variables !

```{r}
df_incumbent_participation_rev <- df_incumbent_participation %>% 
  select(fecha, comuna, candidato_pm, sexo_pm, pacto_pm, pacto_nombre_pm, votos_candidato_pm, votos_candidato_pm_p,
         partido_pm, coalicion_pm, incumbente, sexo_incumbente, pacto_incumbente, pacto_nombre_incumbente, 
         votos_incumbente, votos_incumbente_p, periodos_consecutivos_alcalde, participacion_alcalde_incum,
         reeleccion_alcalde) %>% 
  arrange(comuna, fecha)

#writexl::write_xlsx(df_incumbent_participation_rev, "output/ya basta mamita karina.xlsx")
```



```{r}

library("labelled")

colnames(df_incumbent_participation)

df_final <- df_incumbent_participation %>%
  select(tipo_eleccion, cargo, comuna, cod_com, fecha, periodo_electoral, 
         candidato_pm, sexo_pm, pacto_pm, pacto_nombre_pm, partido_pm, coalicion_pm, votos_candidato_pm, # Primera mayoría
         votos_candidato_pm_p,
         candidato_sm, sexo_sm, pacto_sm, pacto_nombre_sm, partido_sm, coalicion_sm, votos_candidato_sm, # Segunda mayoría
         votos_candidato_sm_p, 
         incumbente, sexo_incumbente, pacto_incumbente, pacto_nombre_incumbente, partido_incumbente, coalicion_incumbente, # Incumbente
         votos_incumbente, votos_incumbente_p,
         brecha_victoria, brecha_victoria_p, coalicion_electa_p = porc_coalicion_alcalde_electa,
         coalicion_incumbente_p = porc_coalicion_alcalde_incumbente, 
         reeleccion_coalicion_alcalde, reeleccion_alcalde, periodos_consecutivos_coalicion, periodos_consecutivos_alcalde,
         participacion_alcalde_incum, elecciones_no_competitivas,
         votos_nulos, votos_en_blanco, votos_comunales, votos_validos, votos_lista, votos_partido, votos_coalicion,
         coalicion_alianza, coalicion_alianza_p, coalicion_concertacion, coalicion_concertacion_p, coalicion_independientes,
         coalicion_independientes_p, coalicion_otros, coalicion_otros_p, coalicion_frente_amplio, coalicion_frente_amplio_p,
         everything()) %>% 
  labelled::set_variable_labels(
                tipo_eleccion = "Tipo de elección", 
                cargo = "Cargo a elegir", 
                comuna = "Comuna", 
                cod_com = "Código comunal", 
                fecha = "Año de elección", 
                periodo_electoral = "Periodo electoral", 
                candidato_pm = "Nombre de candidato electo", 
                sexo_pm = "Sexo de candidato electo", 
                pacto_pm = "Pacto de candidato electo", 
                pacto_nombre_pm = "Nombre del pacto del candidato electo", 
                partido_pm = "Partido del candidato electo", 
                coalicion_pm = "Coalición del candidato electo", 
                votos_candidato_pm = "Votos del candidato electo",
                votos_candidato_pm_p = "Porcentaje de votos obtenidos por el candidato electo del total de votos válidos",
                candidato_sm = "Candidato con la segunda mayoría", 
                sexo_sm = "Sexo del candidato con la segunda mayoría", 
                pacto_sm = "Pacto del candidato con la segunda mayoría", 
                pacto_nombre_sm = "Nombre del pacto del candidato con la segunda mayoría", 
                partido_sm = "Partido del candidato con la segunda mayoría", 
                coalicion_sm = "Coalición del candidato con la segunda mayoría", 
                votos_candidato_sm = " Votos del candidato con la segunda mayoría",
                votos_candidato_sm_p = "Porcentaje de votos obtenidos por el candidato con la segunda mayoría del total de votos válidos",
                brecha_victoria = "Votos de diferencia entre el candidato electo y candidato con la segunda mayoría", 
                brecha_victoria_p = "Porcentaje de diferencia entre el candidato electo y candidato con la segunda mayoría", 
                coalicion_electa_p = "Porcentaje de votos obtenido por la coalición electa",
                coalicion_incumbente_p = "Porcentaje de votos obtenido por la coalición incumbente",
                reeleccion_coalicion_alcalde = "Dummy reelección de la coalición incumbente", 
                reeleccion_alcalde = "Dummy reelección del alcalde incumbente", 
                periodos_consecutivos_coalicion = "Conteo de periodos consecutivos de la coalición incumbente", 
                periodos_consecutivos_alcalde = "Conteo de periodos consecutivos del alcalde incumbente",
                coalicion_incumbente = "Coalición incumbente", 
                participacion_alcalde_incum = "Participación del alcalde incumbente en las elecciones", 
                elecciones_no_competitivas = "Elecciones competitivas",
                votos_nulos = "Votos nulos", 
                votos_en_blanco = "Votos en blanco", 
                votos_comunales = "Total de votos comunales", 
                votos_validos = "Votos válidos", 
                votos_lista = "Votos obtenidos por la lista", 
                votos_partido = "Votos obtenidos por el partido", 
                votos_coalicion = "Votos obtenidos por la coalición",
                coalicion_alianza = "Votos obtenidos por la coalición Alianza", 
                coalicion_alianza_p = "Porcentaje de votos obtenidos por la coalición Alianza",
                coalicion_concertacion = "Votos obtenidos por la coalición Concertación",  
                coalicion_concertacion_p = "Porcentaje de votos obtenidos por la coalición Concertación",
                coalicion_independientes = "Votos obtenidos por independientes fuera de pacto", 
                coalicion_independientes_p = "Porcentaje de votos obtenidos por independientes fuera de pacto",
                coalicion_otros ="Votos obtenidos por otras coaliciones no alianza ni concertación", 
                coalicion_otros_p = "Porcentaje de votos obtenidos por otras coaliciones no alianza ni concertación",
                coalicion_frente_amplio = "Votos obtenidos por la coalición Frente Amplio", 
                coalicion_frente_amplio_p = "Porcentaje de votos obtenidos por la coalición Frente Amplio")

```

Agregar la diferencia de votos de el primero con el segundo:

```{r}
df_final <- df_final %>% 
  group_by(comuna) %>% 
  mutate(diferencia_c_concertacion_p = coalicion_concertacion_p-lag(coalicion_concertacion_p),
         diferencia_c_alianza_p = coalicion_alianza_p-lag(coalicion_alianza_p),
         diferencia_c_otros_p = coalicion_otros_p-lag(coalicion_otros_p),
         diferencia_c_independientes_p = coalicion_independientes_p-lag(coalicion_independientes_p),
         diferencia_c_frente_amplio_p = coalicion_frente_amplio_p-lag(coalicion_frente_amplio_p)) %>% 
  mutate(diferencia_c_incumbente_p = case_when(
    coalicion_incumbente == "concertacion" ~ diferencia_c_concertacion_p,
    coalicion_incumbente == "alianza" ~ diferencia_c_alianza_p,
    coalicion_incumbente == "independientes" ~ diferencia_c_independientes_p,
    coalicion_incumbente == "otros" ~ diferencia_c_otros_p
  ))

df_final <- df_final %>% 
  labelled::set_variable_labels(
    diferencia_c_concertacion_p = "Diferencia del porcentaje total de votos obtenidos por la coalición Concertación en el tiempo t y t-1",
    diferencia_c_alianza_p = "Diferencia del porcentaje total de votos obtenidos por la coalición Alianza en el tiempo t y t-1",
    diferencia_c_otros_p = "Diferencia del porcentaje total de votos obtenidos por otras coaliciones en el tiempo t y t-1",
    diferencia_c_independientes_p = "Diferencia del porcentaje total de votos obtenidos por independientes sin coalición en el tiempo t y t-1",
    diferencia_c_frente_amplio_p = "Diferencia del porcentaje total de votos obtenidos por la coalición Frente Amplio en el tiempo t y t-1",
    diferencia_c_incumbente_p = "Diferencia del porcentaje total de votos obtenidos por la coalición incumbente en la comuna")

df_final <- df_final %>% 
  labelled::set_variable_labels(
    otros_n_candidatos = "Número de candidatos pertenecientes a otras coaliciones",
    alianza_n_candidatos = "Número de candidatos de la Alianza",
    concertacion_n_candidatos = "Número de candidatos de la Concertación",
    independientes_n_candidatos = "Número de candidatos independientes sin coalición",
    frente_amplio_n_candidatos = "Número de candidatos del Frente Amplio",
    votos_coalicion_p = "Porcentaje de votos de la coalición electa",
    votos_coalicion_incumbente_p = "Porcentaje de votos obtenido por la coalición incumbente",
    enp_candidatos = "Número efectivo de candidatos",
    enp_coalicion = "Número efectivo de coaliciones",
    enp_partido = "Número efectivo de partido",
    enp_pacto = "Número efectivo de pactos",
    candidatos_n = "Número total de candidatos",
    coaliciones_n = "Número total de coaliciones",
    partidos_n = "Número total de partidos",
    incumbente = "Nombre del incumbente",
    sexo_incumbente = "Sexo del incumbente",
    pacto_incumbente = "Pacto del incumbente",
    pacto_nombre_incumbente = "Nombre del pacto del incumbente",
    partido_incumbente = "Partido del incumbente",
    votos_incumbente = "Total de votos obtenidos por el incumbente",
    votos_incumbente_p = "Porcentaje de votos obtenidos por el incumbente del total de votos válidos",
    coalicion_municipal_incumbente = "Coalición incumbente de la comuna")

```

```{r}
sjPlot::view_df(df_final, show.labels = T)
```

Guardamos la base de datos:

```{r}
library(haven)

write_dta(df_final, "data/231130 - Elecciones municipales.dta")
write_rds(df_final, "data/231130 - Elecciones municipales.rds")
write_sav(df_final, "data/231130 - Elecciones municipales.sav")

writexl::write_xlsx(df_final, "data/231130 - Elecciones municipales.xlsx")
```
